requirements.txt
anthropic
gradio
mcp
python-dotenv
